Avoid long-living gRPC connections moving the dialing from commands initialization to workers
This change is made after observing a problematic behavior on a Kubernetes
deployment of the whole system. Perhaps due to hidden limits or
misconfiguration, having too many constantly open connections was causing
a general inability of accepting new connection at all, upon a certain
limit.
While it seems not to be a WhatsNew-specific issue, we prefer to keep
it less "aggressive", and open gRPC connections only when needed during
the jobs' execution.